requirements.txt
Django==4.2.13
djangorestframework==3.15.1
django-extensions==3.2.3
django-filter==24.2